Added functionality to display runners on map from database.
Runners are now displayed on map from table race_events.
Created a controller only for populating the database and testing
purposes, ideally we won’t need this in the final version.

TODO:
Need to assert valid lat/long positions that are inserted into table.
Need to fix issue that page needs to be refreshed in order to display
map properly.
Need to add functionality to refresh map with updated coordinates.
Need to figure out how we are going to store desired bibs that want to
be viewed.