Added Robustness and Flexibility to CSV File Types Allowed
The Racers table now includes first and last names of racers. In addition, CSV files with extra headers may now be used. Regular Expressions check that the required field (or variants of them) are available. The CSV file is then parsed, and the necessary fields are written to the database.